feat: :sparkles: More functions and a rework
Now it can't have multiple sessions on the same device, because it only stores a token at a time. Also, decided to add a snackbar for the Login failed message, instead of the view. Plus, now when the login is succesful, the other view shows the username gotten on the payload, and if the user is already logged in, the mock view is launched when first started the app.